Support MSVC 7 (#342)
* msvc 7

* kill long long

* remove sprintf usage

* version

+++ compat/stdio/printf.c
@@ -237,8 +237,8 @@ typedef unsigned int printf_flags_t;
 typedef uint8_t numeric_base_t;
 
 #if PRINTF_SUPPORT_LONG_LONG
-typedef unsigned long long printf_unsigned_value_t;
-typedef long long          printf_signed_value_t;
+typedef uint64_t printf_unsigned_value_t;
+typedef int64_t  printf_signed_value_t;
 #else
 typedef unsigned long printf_unsigned_value_t;
 typedef long          printf_signed_value_t;
@@ -1390,7 +1390,7 @@ static inline void format_string_loop(output_gadget_t* output, const char* forma
 
           if (flags & FLAGS_LONG_LONG) {
 #if PRINTF_SUPPORT_LONG_LONG
-            const long long value = va_arg(args, long long);
+            const int64_t value = va_arg(args, int64_t);
             print_integer(output, ABS_FOR_PRINTING(value), value < 0, base, precision, width, flags);
 #endif
           }
@@ -1419,7 +1419,7 @@ static inline void format_string_loop(output_gadget_t* output, const char* forma
 
           if (flags & FLAGS_LONG_LONG) {
 #if PRINTF_SUPPORT_LONG_LONG
-            print_integer(output, (printf_unsigned_value_t) va_arg(args, unsigned long long), false, base, precision, width, flags);
+            print_integer(output, (printf_unsigned_value_t) va_arg(args, uint64_t), false, base, precision, width, flags);
 #endif
           }
           else if (flags & FLAGS_LONG) {
@@ -1540,7 +1540,7 @@ static inline void format_string_loop(output_gadget_t* output, const char* forma
         else if  (flags & FLAGS_SHORT)     *(va_arg(args, short*))     = (short) output->pos;
         else if  (flags & FLAGS_LONG)      *(va_arg(args, long*))      = (long) output->pos;
 #if PRINTF_SUPPORT_LONG_LONG
-        else if  (flags & FLAGS_LONG_LONG) *(va_arg(args, long long*)) = (long long int) output->pos;
+        else if  (flags & FLAGS_LONG_LONG) *(va_arg(args, int64_t*))   = (int64_t) output->pos;
 #endif /* PRINTF_SUPPORT_LONG_LONG */
         else                               *(va_arg(args, int*))       = (int) output->pos;
         format++;
